Finalise tests
